lambda/db_bootstrap/pyproject.toml
[project]
name = "kubrick_db_bootstrap"
version = "0.1.0"
description = "Lambda Function to bootstrap database"
requires-python = ">=3.13"
dependencies = ["psycopg2-binary>=2.9.10"]

[tool.pyright]
extraPaths = ["../layers/vector_database_layer/", "../layers/config_layer/"]
